+++ src/controllers/UserController.js
@@ -1,4 +1,4 @@
-const { User, Property } = require('../models')
+const { User } = require('../models')
 const deleteFile = require('../utils/deleteFile')
 
 module.exports = {
@@ -29,9 +29,9 @@ module.exports = {
   },
 
   list: async (req, res) => {
-    const { user_id: id } = req.params
+    const { user_id } = req
 
-    User.findByPk(id)
+    User.findByPk(user_id)
       .then((user) => {
         if (!user) {
           return res.status(400).json({
@@ -50,80 +50,36 @@ module.exports = {
       })
   },
 
-  list_properties: async (req, res) => {
-    const { user_id: id } = req.params
-
-    User.findByPk(id, { include: { association: 'properties' } })
-      .then((user) => {
-        if (!user) {
-          return res.status(400).json({
-            cod: 400,
-            message: 'Não conseguimos listar as propriedades do usuário! Por favor, verifique os dados fornecidos e tente novamente.'
-          })
-        }
-
-        return res.json(user.properties)
-      })
-      .catch((err) => {
-        return res.status(500).json({
-          cod: 500,
-          msg: 'Ocorreu um erro inesperado ao listar as propriedades do usuário. Por favor, tentar novamente.'
-        })
-      })
-  },
-
-  list_rentals: async (req, res) => {
-    const { user_id: id } = req.params
-
-    User.findByPk(id, { include: { association: 'rentals' } })
-      .then((user) => {
-        if (!user) {
-          return res.status(400).json({
-            cod: 400,
-            message: 'Não conseguimos listar os aluguéis do usuário! Por favor, verifique os dados fornecidos e tente novamente.'
-          })
-        }
-
-        return res.json(user.rentals)
-      })
-      .catch((err) => {
-        return res.status(500).json({
-          cod: 500,
-          msg: 'Ocorreu um erro inesperado ao listar os aluguéis do usuário. Por favor, tentar novamente.'
-        })
-      })
-  },
-
-  list_favorites: async (req, res) => {
-    const { user_id: id } = req.params
-
-    User.findByPk(id, { include: { association: 'favorites' } })
-      .then((user) => {
-        if (!user) {
-          return res.status(400).json({
-            cod: 400,
-            message: 'Não conseguimos listar os favoritos do usuário! Por favor, verifique os dados fornecidos e tente novamente.'
-          })
-        }
-
-        return res.json(user.favorites)
-      })
-      .catch((err) => {
-        return res.status(500).json({
-          cod: 500,
-          msg: 'Ocorreu um erro inesperado ao listar as propriedades favoritadas pelo usuário. Por favor, tentar novamente.'
-        })
-      })
-  },
-
   update: async (req, res) => {
-    const { user_id: id } = req.params
+    const { user_id } = req
     const { name, cpf, phone } = req.body
+
+    User.update({ name, cpf, phone }, { where: { id: user_id } })
+      .then(([updated]) => {
+        if (!updated) {
+          return res.status(400).json({
+            cod: 400,
+            message: 'Os dados fornecidos são inválidos. Por favor, verifique os dados enviados e tente novamente.'
+          })
+        }
+
+        return res.status(204).json()
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          cod: 500,
+          msg: 'Ocorreu um erro inesperado ao atualizar o usuário. Por favor, tentar novamente.'
+        })
+      })
+  },
+
+  update_image: async (req, res) => {
+    const { user_id } = req
     const { key: avatar } = req.file || { key: undefined }
 
-    const user = await User.findByPk(id)
+    const user = await User.findByPk(user_id)
 
-    User.update({ name, cpf, phone, avatar }, { where: { id } })
+    User.update({ avatar }, { where: { id: user_id } })
       .then(([updated]) => {
         if (!updated) {
           return res.status(400).json({
@@ -145,17 +101,17 @@ module.exports = {
       .catch((err) => {
         return res.status(500).json({
           cod: 500,
-          msg: 'Ocorreu um erro inesperado ao atualizar o usuário. Por favor, tentar novamente.'
+          msg: 'Ocorreu um erro inesperado ao atualizar a imagem do usuário. Por favor, tentar novamente.'
         })
       })
   },
 
   delete: async (req, res) => {
-    const { user_id: id } = req.params
+    const { user_id } = req
 
-    const user = await User.findByPk(id)
+    const user = await User.findByPk(user_id)
 
-    User.destroy({ where: { id } })
+    User.destroy({ where: { id: user_id } })
       .then((deleted) => {
         if (!deleted) {
           return res.status(400).json({
@@ -174,84 +130,6 @@ module.exports = {
         return res.status(500).json({
           cod: 500,
           msg: 'Ocorreu um erro inesperado ao apagar o usuário. Por favor, tentar novamente.'
-        })
-      })
-  },
-
-  add_favorite: async (req, res) => {
-    const { user_id } = req.params
-    const { property_id } = req.body
-
-    const user = await User.findByPk(user_id)
-    if (!user) {
-      return res.status(400).json({
-        cod: 400,
-        msg: 'Este usuário não existe na base de dados.'
-      })
-    }
-
-    const property = await Property.findByPk(property_id)
-    if (!property) {
-      return res.status(400).json({
-        cod: 400,
-        msg: 'Este imóvel não existe na base de dados.'
-      })
-    }
-
-    user.addFavorite(property)
-      .then((favorite) => {
-        if (!favorite) {
-          return res.status(400).json({
-            cod: 400,
-            msg: 'Ocorreu um erro ao adicionar este imóvel como favorito.'
-          })
-        }
-
-        return res.status(204).json()
-      })
-      .catch((err) => {
-        return res.status(500).json({
-          cod: 500,
-          msg: 'Ocorreu um erro inesperado ao adicionar o imóvel como favorito. Por favor, tentar novamente.'
-        })
-      })
-  },
-
-  remove_favorite: async (req, res) => {
-    const { user_id } = req.params
-    const { property_id } = req.body
-
-    const user = await User.findByPk(user_id)
-    if (!user) {
-      return res.status(400).json({
-        cod: 400,
-        msg: 'Este usuário não existe na base de dados.'
-      })
-    }
-
-    const property = await Property.findByPk(property_id)
-    if (!property) {
-      return res.status(400).json({
-        cod: 400,
-        msg: 'Este imóvel não existe na base de dados.'
-      })
-    }
-
-    user.removeFavorite(property)
-      .then((deleted) => {
-        if (!deleted) {
-          return res.status(400).json({
-            cod: 400,
-            msg: 'Ocorreu um erro ao apagar este imóvel dos favoritos.'
-          })
-        }
-
-        return res.status(204).json()
-      })
-      .catch((err) => {
-        return res.status(500).json({
-          cod: 500,
-          msg: 'Ocorreu um erro inesperado ao apagar o imóvel dos favoritos. Por favor, tentar novamente.'
         })
       })
   }
